+++ src/components/Contacts.vue
@@ -67,7 +67,7 @@ let contacts: ComputedRef<Contact[]> = computed(() => {
 <template>
   <section class="mb-5 p-5 border-4 border-gray-500">
     <SectionTitle anchor="contacts" title="Contacts" />
-    <div class="flex mb-2">
+    <div class="flex">
       <form class="w-1/2 pr-2 flex flex-col">
         <label
           v-for="individual in individuals"